+++ changelog.md
@@ -1,2 +1,5 @@
+**Requires [pv-addon-lavaplayer-lib 1.0.11+](https://modrinth.com/plugin/pv-addon-lavaplayer-lib/version/1.0.11) to work**
+
 Changelog:
 - Fixed an issue where inserting a disc with shift-click caused it to start playing on 1.21
+- Reintroduced YouTube authorization